demo modify

+++ demo/timer/main.cpp
@@ -8,10 +8,10 @@ int main()
     eb.Init();
 
     eb.AddTimer(1000, true, [](EventHandle _handle) {
-        Logger(LM_INFO) << "On Timer:" << String::ToString(reinterpret_cast<uint64_t>(_handle));
+        Logger(LM_INFO) << "On Timer:" << String::ToString(reinterpret_cast<unsigned long long>(_handle));
         });
     eb.AddSignal(SIGINT, false, [&](int _signal, EventHandle _handle) {
-        Logger(LM_INFO) << "On Signal:" << _signal << " trigger. " << String::ToString(reinterpret_cast<uint64_t>(_handle));
+        Logger(LM_INFO) << "On Signal:" << _signal << " trigger. " << String::ToString(reinterpret_cast<unsigned long long>(_handle));
         eb.EventLoopBreak();
         });
